Wep Api

+++ Entities/Concrete/ChatLevel.cs
@@ -4,7 +4,7 @@ namespace Entities.Concrete
 {
     public class ChatLevel : IEntity
     {
-        public int ChatLevelId { get; set; }
+        public byte ChatLevelId { get; set; }
         public string Level { get; set; }
     }
