+++ Entities/Concrete/Match.cs
@@ -7,7 +7,7 @@ namespace Entities.Concrete
         public int MatchId { get; set; }
         public int MatchUserId { get; set; }
         public int MatchedUserId { get; set; }
-        public int ChatLevelId { get; set; }
+        public byte ChatLevelId { get; set; }
     }
 
 }
